feat(all-posts): 「烏は主を選ばない」記事追加とdeploy.sh修正
- all-posts.orgにアニメ「烏は主を選ばない」の投稿を追加
- カバー画像(mountain.jpg, IMG_0256.HEIC)を追加
- deploy.shのカバー画像コピー先にファイル名(cover.jpg)を明示するよう修正
- 新規・既存投稿のリサイズ済み生成画像を追加

+++ deploy.sh
@@ -45,10 +45,10 @@ cd ${hugo}/content/post && \
     do
         d=$(basename $f .md)
         if [ -d ${d} ]; then
-            test -f ${d}/cover.* || cp ${hugo}/assets/img/cover/${m}.jpg ${d}
+            test -f ${d}/cover.* || cp ${hugo}/assets/img/cover/${m}.jpg ${d}/cover.jpg
         else
             mkdir ${d}
-            cp ${hugo}/assets/img/cover/${m}.jpg ${d}
+            cp ${hugo}/assets/img/cover/${m}.jpg ${d}/cover.jpg
         fi
     done
